requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.1.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==2.1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
dm-tree==0.1.7
drjit==0.4.1
entrypoints==0.4
executing==0.8.3
flatbuffers==23.3.3
fonttools==4.34.4
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.6.0
idna==2.8
importlib-metadata==4.11.4
importlib-resources==5.12.0
ipydatawidgets==4.3.2
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.5
jax==0.4.7
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.6
keras==2.11.0
keras-cv==0.2.7
keras-flops==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
mitsuba==3.2.1
ml-dtypes==0.0.4
model-profiler==1.1.8
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyGObject==3.36.0
pyparsing==3.0.9
# python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
pythreejs==2.4.2
pytz==2022.1
pyzmq==23.2.0
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scipy==1.10.1
seaborn==0.11.2
sionna==0.14.0
six==1.14.0
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-compression==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-probability==0.17.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
traittypes==0.2.1
typeguard==2.13.3
typing-extensions==4.2.0
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==2.1.2
widgetsnbextension==4.0.6
wrapt==1.14.1
zipp==3.8.0
